ai and health bar integration
Integrated the health and stamin uis to test damage on the player and if the ui's would update. Updated the health script to  script for the slider value and include the stamina bar, along with having the stamina bar regenerate very slow in real time. Added a health bar for the enemy ai, and made an ability damage script to attach to abilities. This script will do damage based on the values inputted and subtract from the enemy's current health bar.