Merging scripts and working timer
I made the timer functional with it have milliseconds, seconds, minutes, and hours. When you press the pause button the timer stops, and when you buy an item the timer increases by the amount set for the specific item. I also merged the menu scaler and the menu animation scripts to make it more compact with scripts.