Fixed database scoring system
altered the timer script to also have a count that counts up every second. The counter is then subtracted from an incredibly large number to get the current score. The save data script was also update to handle the current score and have it reflect in the inspector field for the database.