update wind boss ai
Updated the wind boss ai to pull the play to a desired location, and prevent the player from moving beyond the limit set for the x-axis but still allow movement along the z axis. The movement is restricted with an active force being applied to the player. The boss will also move to a fixed point in the arena before pulling the player.