Update ui trgger and scene trigger
Add trigers for both the score and leaderboard uis to show after the ai game object is destroyed if the death flag is set. Also added a script to control scene changes. This is for when the player defeats a boss to get transported to the rest area and then interact with a game object to begin the next fight.