update multiplayer camera and inputs
The camera will now automatically assign to the player object owner which will allow the camera to follow the player around the arena in a multiplayer setting. The inputs of the player are now exclusive to the owner, allowing for multiple players to be present without all of them being forced to use a player input from one player.