fixed game reset after clicking new game
the game now properly resets the timer, and death counter to start the game as if  it is your first time through. Fixed an issue when player dies in tutorial before beating the tutorial boss, the death counter remained at zero. Updated sounds for attacks, and added sounds for when items are used.