updated game save system
the system now checks if the game has been completed. If it has, then the player is unable to continue from last save. This forces the player to start a new game. after starting a new game, when the player saves the game the game checker bool will then be reset to allow to player to interact with the continue button.